Use Timer for startup delay, `advanceIfElapsed`
The timer-based startup delay simplifies startup code. Also, `advanceIfElapsed` is I think technically more reliable than `hasElapsed` and `reset` because it only resets the duration, not entirely (so it's not off by a few milliseconds depending on how long periodic takes to run). It's not necessary but I think it's cleaner.